Minor stylistic changes to search.[(c)|(h)]

+++ searching/search.h
@@ -2,8 +2,8 @@
 
 #include "../arrays/array.h"
 
-int linear_search(int *array, int length, int element);
+int linear_search(int *array, size_t length, int element);
 
-int linear_search_rec(int *array, int length, int element);
+int linear_search_rec(int *array, size_t length, int element);
 
 int binary_search(int *array, int element, size_t start, size_t end);
